Removed margin constraints in App.css; created separate Sidebar subdirectory and split files into implementation and styling; changed styling of Footer to get closer to spec; added margins to Navbar to get closer to design file
