+++ frontend/src/App.css
@@ -4,8 +4,6 @@
 
 .App {
   text-align: center;
-  max-width: 1020px;
-  margin: auto;
 }
 
 .App-logo {
