+++ frontend/src/App.tsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import { RouteProps, Route, Switch } from 'react-router';
 import { LandingPage, Houses, Login, PostRegister } from './pages/index';
+import { Sidebar } from './components/shared_components/index';
 import './App.css';
 
 
 const App = (props: RouteProps) => {
   return (
     <div className='App'>
+      <Sidebar />
       <Switch>
         <Route exact path='/' component={LandingPage} />
         <Route path='/Login' component={Login} />
