+++ frontend/src/components/shared_components/Container.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 
 const StyledContainer = styled('div')`
-  font-family: roboto;
+  font-family: Roboto;
   position: relative;
   width: 70%;
   margin: 0 auto;
