+++ frontend/src/components/shared_components/index.ts
@@ -1,5 +1,6 @@
 import Button from './Button';
 import SpecialButton from './SpecialButton';
 import Container from './Container';
+import Sidebar from './Sidebar/Sidebar';
 
-export { Button, SpecialButton, Container };
+export { Button, SpecialButton, Container, Sidebar };
